requirements.txt
google-cloud-storage == 1.43.0
numpy == 1.21.2
pandas == 1.3.3
tensorflow == 2.7.0
tensorflow-text == 2.7.3
